Implement uptest() and lotest()
Since ROT13 specifies that it only impacts alphabetic characters, it is
necessary to test [buff] to ensure that it is within the ranges 41h
to 5Ah or 61h to 7Ah.

The uptest() tests to see if [buff] is in that former range. If
yes, it properly ROT13 encodes it and jumps to write(). If
not, it either writes it unmodified or passes it to the lotest().

The lotest() tests to see if [buff] is in that latter range. Since
uptest() has already tested [buff], anything outside that range is
passed to write(). Anything within that range will be encoded, THEN
passed to write().

Both ranges are covered, and anything outside them is written out
unaltered.

As yet, write() is not implemented.